# Solarica/connector/api-contract/openapi.yaml
openapi: "3.1.0"
info:
  title: Solarica Connector API
  version: "1.0.0"
  description: |
    Shared REST API contract implemented by both the Python and .NET connectors.
    The connector runs locally on port 8765 and is called by the Solarica web app.

    **Rules:**
    - Both connectors MUST implement every endpoint listed here.
    - All field names use camelCase in JSON.
    - The `/health` response `runtime` field identifies which connector is active.
    - Both connectors listen on `http://127.0.0.1:8765` by default.
servers:
  - url: http://127.0.0.1:8765
    description: Local connector (default port)
...
paths:
...
  /api/device/connect:
    post:
      tags: [device]
      summary: Connect to a specific port
      operationId: connectPort
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectRequest"
      responses:
        "200":
          description: Updated device status after connect attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceStatus"
...
  /api/import/start:
    post:
      tags: [import]
      summary: Pull measurements from the active driver into local SQLite cache
      operationId: startImport
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportStartResult"
        "501":
          description: Driver not implemented (e.g. serial protocol pending)
...
  /api/measurements/{id}:
    get:
      tags: [measurements]
      summary: Get a single measurement with full I-V curve data
      operationId: getMeasurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Measurement detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
        "404":
          description: Not found
...
components:
  schemas:
...
    ConnectRequest:
      type: object
      required: [port]
      properties:
        port:
          type: string
          example: COM3
    DeviceStatus:
      type: object
      required: [connected, mode]
      properties:
        connected:
          type: boolean
        mode:
          type: string
          enum: [mock, serial, vendor_export]
        port:
          type: string
          nullable: true
        deviceModel:
          type: string
          nullable: true
        deviceSerial:
          type: string
          nullable: true
        firmwareVersion:
          type: string
          nullable: true
        transferModeRequired:
          type: boolean
          default: true
        transferModeDetected:
          type: boolean
          default: false
        lastError:
          type: string
          nullable: true
    ImportStartResult:
      type: object
      required: [ok, imported]
      properties:
        ok:
          type: boolean
        imported:
          type: integer
          example: 12
...
    CurvePoint:
      type: object
      required: [pointIndex, voltageV, currentA]
      properties:
        pointIndex:
          type: integer
        voltageV:
          type: number
          format: double
        currentA:
          type: number
          format: double
    Measurement:
      type: object
      required: [id, measuredAt, syncStatus]
      properties:
        id:
          type: string
        externalMeasurementKey:
          type: string
          nullable: true
        measuredAt:
          type: string
          format: date-time
        customer:
          type: string
          nullable: true
        installation:
          type: string
          nullable: true
        stringNo:
          type: string
          nullable: true
        moduleType:
          type: string
          nullable: true
        moduleReference:
          type: string
          nullable: true
        modulesSeries:
          type: integer
          nullable: true
        modulesParallel:
          type: integer
          nullable: true
        nominalPowerW:
          type: number
          nullable: true
        ppkWp:
          type: number
          nullable: true
        rsOhm:
          type: number
          nullable: true
        rpOhm:
          type: number
          nullable: true
        vocV:
          type: number
          nullable: true
        iscA:
          type: number
          nullable: true
        vpmaxV:
          type: number
          nullable: true
        ipmaxA:
          type: number
          nullable: true
        ffPercent:
          type: number
          nullable: true
        sweepDurationMs:
          type: number
          nullable: true
        irradianceWM2:
          type: number
          nullable: true
        sensorTempC:
          type: number
          nullable: true
        moduleTempC:
          type: number
          nullable: true
        irradianceSensorType:
          type: string
          nullable: true
        irradianceSensorSerial:
          type: string
          nullable: true
        importSource:
          type: string
          nullable: true
        syncStatus:
          type: string
          enum: [unsynced, synced, error]
          default: unsynced
        notes:
          type: string
          nullable: true
        curvePoints:
          type: array
          items:
            $ref: "#/components/schemas/CurvePoint"
...